+++ src/snowflake/snowpark/modin/plugin/compiler/snowflake_query_compiler.py
@@ -8884,7 +8884,20 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         # materially slow down CI or individual groupby.apply() calls.
         # TODO(SNOW-1345395): Investigate why and to what extent the cache_result
         # is useful.
-        ordered_dataframe = cache_result(udtf_dataframe)
+        try:
+            ordered_dataframe = cache_result(udtf_dataframe)
+        except SnowparkSQLException as e:
+            if "No module named 'snowflake'" in str(
+                e
+            ) or "Modin is not installed" in str(e):
+                raise SnowparkSQLException(
+                    "modin.pandas cannot be referenced within a Snowpark pandas apply() function. "
+                    "You can only use native pandas inside apply(). Please check developer guide for details "
+                    "https://docs.snowflake.com/developer-guide/snowpark/python/pandas-on-snowflake#limitations."
+                )
+            else:
+                # retry the try-block logic
+                ordered_dataframe = cache_result(udtf_dataframe)
 
         # After applying the udtf, the underlying Snowpark DataFrame becomes
         # -------------------------------------------------------------------------------------------